backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose
passlib[argon2]
langgraph